+++ bidsme/process.py
@@ -32,7 +32,7 @@ from bidsme import exceptions
 from bidsme import plugins
 from bidsme import Modules
 
-from bidsme.tools import selector
+from bidsme.tools import type_selector
 from bidsme.tools import paths
 from bidsme.tools import tools
 
@@ -41,6 +41,7 @@ from bidsme.bidsMeta import BidsSession
 from bidsme.bidsMeta import BidsTable
 
 logger = logging.getLogger(__name__)
+selector = type_selector()
 
 
 def coin(destination: str,